Export/Import memory, token limit, prompt limit, notes update

+++ routes.py
@@ -3,12 +3,12 @@ import json
 import os
 import signal
 from fastapi import APIRouter, HTTPException, BackgroundTasks, File, Request, UploadFile, WebSocket, WebSocketDisconnect, Form, Request
-from fastapi.responses import FileResponse, HTMLResponse, Response
+from fastapi.responses import FileResponse, HTMLResponse, Response, JSONResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from utils import process_message, OpenAIResponser, AudioProcessor, AddonManager, SettingsManager, BrainProcessor
+from utils import process_message, OpenAIResponser, AudioProcessor, AddonManager, SettingsManager, BrainProcessor, MessageParser
 from authentication import Authentication
-from classes import User, UserCheckToken, UserName, editSettings, userMessage
+from classes import User, UserCheckToken, UserName, editSettings, userMessage, noTokenMessage
 import shutil
 from database import Database
 from memory import export_memory_to_file, import_file_to_memory, wipe_all_memories, stop_database
@@ -46,18 +46,6 @@ async def read_root(request: Request):
             return templates.TemplateResponse("local_index.html", {"request": request, "version": version})
         except FileNotFoundError:
             raise HTTPException(status_code=404, detail="Item not found")
-  
-@router.get("/{file_path:path}")
-async def read_file(file_path: str):
-    try:
-        if os.path.exists(f'static/{file_path}'):
-            response = FileResponse(f'static/{file_path}')
-            response.headers["Cache-Control"] = "public, max-age=86400"
-            return response
-        else:
-            raise HTTPException(status_code=404, detail="Item not found")
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="Item not found")
 
 def get_token(request: Request):
     return request.cookies.get('session_token')
@@ -296,16 +284,29 @@ def get_token_usage(username):
     db.open()
     # get the user_id for the given username
     db.cursor.execute(f"SELECT id FROM users WHERE username = '{username}'")
-    user_id = db.cursor.fetchone()[0]
+    user_id = db.cursor.fetchone()
+    if user_id is None:
+        db.close()
+        return 0, 0
+
+    user_id = user_id[0]
 
     # get the token usage from the statistics table
     db.cursor.execute(f"SELECT total_tokens_used, prompt_tokens, completion_tokens FROM statistics WHERE user_id = {user_id}")
     result = db.cursor.fetchone()
     db.close()
-    prompt_cost = round(result[1] * 0.03 / 1000, 3)
-    completion_cost = round(result[2] * 0.06 / 1000, 3)
+    if result is None:
+        return 0, 0 
+
+    prompt_tokens_used = result[1] if result[1] is not None else 0
+    completion_tokens_used = result[2] if result[2] is not None else 0
+    prompt_cost = round(prompt_tokens_used * 0.03 / 1000, 3)
+    completion_cost = round(completion_tokens_used * 0.06 / 1000, 3)
     total_cost = round(prompt_cost + completion_cost, 3)
     return result[0], total_cost
+
+def count_tokens(message):
+    return MessageParser.num_tokens_from_string(message)
 
 # update settings route
 @router.post("/update_settings/",
@@ -416,6 +417,8 @@ async def handle_message(request: Request, message: userMessage, background_task
     success = auth.check_token(message.username, session_token)
     if not success:
         raise HTTPException(status_code=401, detail="Token is invalid")
+    if count_tokens(message.prompt) > 1000:
+        raise HTTPException(status_code=400, detail="Prompt is too long")
     return await process_message(message.prompt, message.username, background_tasks, users_dir)
 
 # openAI response route without modules
@@ -523,7 +526,8 @@ async def handle_generate_audio(request: Request, message: userMessage, backgrou
     success = auth.check_token(username, session_token)
     if not success:
         raise HTTPException(status_code=401, detail="Token is invalid")
-    
+    if count_tokens(message.prompt) > 1000:
+        raise HTTPException(status_code=400, detail="Prompt is too long")    
     audio_path, audio = await AudioProcessor.generate_audio(message.prompt, username, users_dir)
     return FileResponse(audio_path, media_type="audio/wav")
 
@@ -649,38 +653,51 @@ async def handle_delete_data(request: Request, user: UserName):
             },
             },)
 async def handle_upload_data(request: Request, username: str = Form(...), data_file: UploadFile = File(...)):
-    print(username)
-    print(data_file.filename)
     session_token = request.cookies.get("session_token")
     auth = Authentication()
     success = auth.check_token(username, session_token)
     if not success:
         raise HTTPException(status_code=401, detail="Token is invalid")
-    
+
     # Save the uploaded file to the user's directory
     file_path = os.path.join(users_dir, username, data_file.filename)
     with open(file_path, "wb") as buffer:
         shutil.copyfileobj(data_file.file, buffer)
-    
+
     # Check if the file is a zip file
     if zipfile.is_zipfile(file_path):
         # Extract the zip file
         with zipfile.ZipFile(file_path, 'r') as zip_ref:
             zip_ref.extractall(os.path.join(users_dir, username))
         
+        # Check for 'memory.json', and 'settings.json'
+        if 'memory.json' not in os.listdir(os.path.join(users_dir, username)) or \
+        'settings.json' not in os.listdir(os.path.join(users_dir, username)):
+            raise HTTPException(status_code=400, detail="Zip file is missing required files")
+
+        # Verify 'memory.json' is a valid JSON file
+        try:
+            with open(os.path.join(users_dir, username, 'memory.json')) as f:
+                json.load(f)
+        except json.JSONDecodeError:
+            raise HTTPException(status_code=400, detail="'memory.json' is not a valid JSON file")
+
+        # Verify 'settings.json' is a valid JSON file
+        try:
+            with open(os.path.join(users_dir, username, 'settings.json')) as f:
+                json.load(f)
+        except json.JSONDecodeError:
+            raise HTTPException(status_code=400, detail="'settings.json' is not a valid JSON file")
+
         # Import the data into memory
         import_file_to_memory(path=os.path.join(users_dir, username, 'memory.json'), replace=True, username=username)
-        
+
         # Delete the zip file
         os.remove(file_path)
     else:
-        # Import the data into memory
-        import_file_to_memory(path=file_path, replace=True, username=username)
-        
-        # Delete the .json file
-        os.remove(file_path)
-    
-    return {'message': 'User data uploaded successfully'}
+        raise HTTPException(status_code=400, detail="Wrong file format!")
+
+    return JSONResponse(content={'message': 'User data uploaded successfully'})
 
 @router.post("/abort_button/",
     tags=["Messaging"],
@@ -713,6 +730,33 @@ async def handle_abort_button(request: Request, user: UserName):
         raise HTTPException(status_code=401, detail="Token is invalid")
     # kill the process, this will need a manual restart or a cron job/supervisor/...
     os.kill(os.getpid(), signal.SIGINT)
+
+# no token message route
+@router.post("/notoken_message/",
+    tags=["Messaging"],
+    summary="Send a message without token",
+    description="This endpoint allows the user to send a message to the AI by providing a prompt, username and password. The AI will respond to the prompt.",
+    response_description="Returns the AI's response.",
+    responses={
+        200: {
+            "description": "AI responded successfully",
+            "content": {
+                "application/json": {
+                    "example": {"message": "AI responded successfully"}
+                }
+            },
+        },
+    },
+)
+
+async def handle_notoken_message(message: noTokenMessage):
+    auth = Authentication()
+    session_token = auth.login(message.username, message.password)
+    if not session_token:
+        raise HTTPException(status_code=401, detail="User login failed")
+    if count_tokens(message.prompt) > 1000:
+        raise HTTPException(status_code=400, detail="Prompt is too long")
+    return await process_message(message.prompt, message.username, None, users_dir)
 
 @router.get("/admin/statistics", tags=["Admin"],
             summary="Get statistics",
@@ -787,3 +831,15 @@ async def get_statistics():
         "total_voice_usage": total_voice_usage,
         "user_statistics": user_statistics
     }
+
+@router.get("/{file_path:path}")
+async def read_file(file_path: str):
+    try:
+        if os.path.exists(f'static/{file_path}'):
+            response = FileResponse(f'static/{file_path}')
+            response.headers["Cache-Control"] = "public, max-age=86400"
+            return response
+        else:
+            raise HTTPException(status_code=404, detail="Item not found")
+    except FileNotFoundError:
+        raise HTTPException(status_code=404, detail="Item not found")
